Update link for Udi Dahan's Advanced Distributed Systems Design course
Link replaced as the old link was defunct and going to a squatted landing page.  The new link navigates to the landing page (branded with Udi's company, Particular Software) for the online course and can be directly purchased or trialled from there.

Updated the description to indicate that it's a online course rather than a live course.